docs: add missing information in status bar docs (#12618)
**Motivation**

Recently noticed that docs for `statusBarStyle` and `statusBarHidden`
are not kept in sync between files in `react-native-screens` repo,
`react-naviagtion` in-code docs and `reactnavigation.org/docs`.

I added information about support on Android for `statusBarStyle` and
`statusBarHidden` props and `auto` and `inverted` values not being
supported on Android for `statusBarStyle`.

Similar PR in `react-native-screens`:
https://github.com/software-mansion/react-native-screens/pull/2953,
`react-navigation.github.io`:
https://github.com/react-navigation/react-navigation.github.io/pull/1432.