feat: accept name argument for useRoute (#12820)
**Motivation**

This updates `useRoute` to use a type-safe API.

- The `useRoute` hook now accepts an argument - the name of the route
- The argument to `useRoute` provides auto-completion based on all
routes in the project (using `ReactNavigation.RootParamList`)
- The returned type of `useRoute` is based on the route name argument
- When no argument is passed, `useRoute` returns a union of all routes
in the project

BREAKING CHANGE: `useRoute` no longer accepts a generic to override the
returned type since it's not type-safe. User can do `const route =
useRoute() as SomeType` to make it explicit.


https://github.com/user-attachments/assets/512ab42e-af24-452e-8ef1-99276a3ba17c

NOTE: This doesn't check parent routes. It's being worked on in
https://github.com/react-navigation/react-navigation/pull/12790

**Test plan**

- Added unit tests and type tests