fix: inactive screen visible for one frame (#11315)
In a scenario when we have 3 screens pushed on the stack (let's name
them A, B, C in the nesting order) and we navigate from C to B, the A
screen was attached for about one frame (and it should not be since it
will not be visible after transition), causing problems with libraries
that subscribe to native attach/detach process.

It happened because when the third screen disappears, the length of the
routes becomes smaller, so in our case length becomes `2`,
`activeScreensLimit` is still `1`, so we do go into `else` in line `609`
for screen `A`, which was inactive during the transition. In there, the
`outputValue` is set correctly to `STATE_INACTIVE` for after the
transition, but it is interpolated from `1` to `0`(`STATE_INACTIVE`)
(lines `611-616`), so it gets the `1` value. The transition ends in the
next frame, so it quickly becomes `0` again, but still was attached for
this one frame.

The solution is to keep the information of what was the last state for
each index, and if it was `STATE_INACTIVE` and will be such after the
transition, there is no sense to make it visible. It would be nice to
test it in many scenarios for sure.

@graszka22 can confirm that it fixes a problem with native attach/detach
in their library.

---------

Co-authored-by: Kacper Kapuściak <39658211+kacperkapusciak@users.noreply.github.com>
Co-authored-by: Satyajit Sahoo <satyajit.happy@gmail.com>