feat: add sheetShouldOverflowTopInset prop (#12871)
We unified the behavior of form sheets across Android and iOS in `react-native-screens`. On iOS, form sheets respect the top safe area inset and do not extend underneath the status bar. We’ve updated the Android implementation to match this behavior for consistency across platforms.

To maintain backward compatibility, we introduce a new prop `sheetShouldOverflowTopInset` that allows reverting to the previous behavior where the component could extend beneath the status bar.

**Motivation**

This change enhances visual consistency.

Additionally, it allows to handle the case with moving the FormSheet content above the keyboard and avoiding moving it behind the status bar or display cutout

related PRs:
- https://github.com/software-mansion/react-native-screens/pull/3404
- https://github.com/software-mansion/react-native-screens/pull/3248

**Test plan**

Tested with the example added to `react-native screens` in https://github.com/software-mansion/react-native-screens/pull/3404