fix: react-native-tab-view glitching on initial load (#12627)
**Motivation**

While working for Expensify, we encountered a bug with glitching
material top tabs. After investigation, it turned out that the
`react-native-tab-view` causes it.

It happens on refresh or initial load of the page, when any tab except
the first one is selected.

Disclaimer: In our case, we use the `position` value to animate the top
bar.

The same effect can be reproduced in an example app with material top
tabs. Hover on the example app, it is not as visible as in the
production app due to the performance.

In the example app, it was visible for one or two frames.


https://github.com/user-attachments/assets/a285cce4-333d-4294-854f-a82057820d42

The root cause is:
- `layout.width` determines if we should use a layout that uses the
`position` and `translateX` animated values.
- Both `position` and `translateX` are derived from `panX`
- There is a brief moment where `layout.width` is known, so we use
`position` and `translateX`, but the `panX` isn't set yet and is still
0.

I changed the code to set `panX` within `useLayoutEffect` instead of
`useEffect`

After fixing:


https://github.com/user-attachments/assets/f1eeda37-8861-4be7-8d4f-237a7438446a

**Test plan**

See if the glitching occurs

The change must pass lint, typescript and tests.